Decompile func0f0cf38c

+++ src/include/game/data/data_0160b0.h
@@ -94,7 +94,7 @@ extern u32 var80070e84;
 extern u32 var80070e90;
 extern u32 var80070e9c;
 extern u32 var80070ea8;
-extern u32 var80070ec0;
+extern u8 var80070ec0;
 extern u8 var80070ec4;
 extern f32 g_EyespyMaxHeight;
 extern u32 var80070ed4;
